+++ src/renderer/opengl.rs
@@ -45,6 +45,9 @@ pub struct OpenGl {
     main_programs_with_glyph_texture: [Option<MainProgram>; 11],
     // Same shader programs but with has_glyph_texture being false
     main_programs_without_glyph_texture: [Option<MainProgram>; 11],
+    main_programs_with_glyph_texture: [Option<MainProgram>; 12],
+    // Same shader programs but with has_glyph_texture being false
+    main_programs_without_glyph_texture: [Option<MainProgram>; 12],
     current_program: u8,
     current_program_needs_glyph_texture: bool,
     vert_arr: Option<<glow::Context as glow::HasContext>::VertexArray>,
@@ -194,6 +197,18 @@ impl OpenGl {
                         false,
                     )?)
                 },
+                Some(MainProgram::new(
+                    &context,
+                    antialias,
+                    ShaderType::FillGradientRadial,
+                    with_glyph_texture,
+                )?),
+                Some(MainProgram::new(
+                    &context,
+                    antialias,
+                    ShaderType::FillImageGradientRadial,
+                    with_glyph_texture,
+                )?),
             ])
         };